func-tests: remove setup.py and add classes to entry.py
