+++ functional-tests/fn_bridge_deposit_reclaim.py
@@ -8,7 +8,7 @@ from strata_utils import (
 )
 
 from constants import DEFAULT_ROLLUP_PARAMS, DEFAULT_TAKEBACK_TIMEOUT, UNSPENDABLE_ADDRESS
-from setup import TestStrata
+from entry import TestStrata
 from utils import get_bridge_pubkey, wait_until
 
 # Local constants
