+++ functional-tests/fn_btcio_read_reorg.py
@@ -4,8 +4,7 @@ import flexitest
 from bitcoinlib.services.bitcoind import BitcoindClient
 
 from constants import BLOCK_GENERATION_INTERVAL_SECS, SEQ_SLACK_TIME_SECS
-from entry import BasicEnvConfig
-from setup import TestStrata
+from entry import BasicEnvConfig, TestStrata
 from utils import wait_until
 
 REORG_DEPTH = 3
