+++ functional-tests/fn_client_status.py
@@ -1,6 +1,6 @@
 import flexitest
 
-from setup import TestStrata
+from entry import TestStrata
 from utils import wait_until
