+++ functional-tests/fn_el_balance_transfer.py
@@ -3,7 +3,7 @@ import time
 import flexitest
 from web3 import Web3
 
-from setup import TestStrata
+from entry import TestStrata
 
 
 @flexitest.register
