+++ functional-tests/fn_prover_checkpoint_runner.py
@@ -2,7 +2,7 @@ import time
 
 import flexitest
 
-from setup import TestStrata
+from entry import TestStrata
 
 
 @flexitest.register
