+++ functional-tests/fn_rpc_bridge_msg.py
@@ -2,7 +2,7 @@ import time
 
 import flexitest
 
-from setup import TestStrata
+from entry import TestStrata
 
 WAIT_TIME = 2
