+++ functional-tests/fn_rpc_exec_update.py
@@ -2,7 +2,7 @@ import time
 
 import flexitest
 
-from setup import TestStrata
+from entry import TestStrata
 
 
 @flexitest.register
