+++ functional-tests/fn_sync_from_rpc.py
@@ -2,7 +2,7 @@ import time
 
 import flexitest
 
-from setup import TestStrata
+from entry import TestStrata
 
 
 @flexitest.register
